packages/spark-advisor-gateway/pyproject.toml
[project]
name = "spark-advisor-gateway"
version = "0.1.0"
description = "API Gateway for spark-advisor - REST + NATS task orchestration"
requires-python = ">=3.12"
dependencies = [
    "spark-advisor-models",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "nats-py>=2.9",
    "orjson>=3.9",
    "pyyaml>=6.0",
]

[project.scripts]
spark-advisor-gateway = "spark_advisor_gateway.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spark_advisor_gateway"]

[tool.uv.sources]
spark-advisor-models = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=1.0",
    "pytest-cov>=6.1",
    "httpx>=0.28",
    "mypy>=1.15",
    "ruff>=0.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "TCH", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["spark_advisor_gateway", "spark_advisor_models"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "pydantic_settings.BaseSettings"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["spark_advisor_gateway"]
branch = true

[tool.coverage.report]
show_missing = true
